Miglioramento della rotta getSubInterval (T5) gestita attraverso corretto pattern Controller/Service e aggiunta rotta all'interno dell'API gateway

+++ T4-G18/api/leaderboard/leaderboard.go
@@ -4,7 +4,7 @@ import "strconv"
 
 type Row struct {
 	UserID int64 `json:"userId"`
-	Stat   int64 `json:"statistic"`
+	Stat   int32 `json:"statistic"`
 }
 
 type Leaderboard struct {
